chore: Move cursor context menu
