+++ web_ui/src/pages/create-project/components/pose-template/canvas/cursor-context-menu/cursor-context-menu.test.tsx
