+++ web_ui/src/pages/create-project/components/pose-template/canvas/edge.component.tsx
@@ -9,7 +9,7 @@ import { DOMRefValue } from '@react-types/shared';
 import clsx from 'clsx';
 
 import { KeypointNode, Point } from '../../../../../core/annotations/shapes.interface';
-import { CursorContextMenu } from '../../../../../shared/components/cursor-context-menu/cursor-context-menu.component';
+import { CursorContextMenu } from './cursor-context-menu/cursor-context-menu.component';
 import { HiddenEdge } from './hidden-edge.component';
 
 import classes from './edge.module.scss';
